Primeira tela de dashboard

+++ src/main/java/eventify/api_spring/dto/smartsync/dashboard/RegistroKpiDto.java
@@ -1,5 +1,8 @@
 package eventify.api_spring.dto.smartsync.dashboard;
 
+import java.math.BigDecimal;
+import java.util.List;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,9 +13,9 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RegistroKpiDto {
-    private Integer ultimaSemana;
-    private Integer ultimosTresMeses;
-    private Integer total;
-    private Integer percentualContratantes;
-    private Integer percentualProprietarios;
+    private Long ultimaSemana;
+    private Long ultimosTresMeses;
+    private Long total;
+    private BigDecimal percentualContratantes;
+    private BigDecimal percentualProprietarios;
 }
